new - structure for types add - form for new products
