+++ frontend/src/types/Product.tsx
@@ -1,0 +1,9 @@
+export type Product = {
+    id: number,
+    productId: string,
+    name: string,
+    category: string,
+    quantity: number,
+    status: string
+
+}
